# openapi-action.yaml
openapi: 3.1.0
info:
  title: Second Perspective Super Decision-Hub
  description: Auditable decision orchestration with deterministic evaluation, causal counterfactuals,
    scenario stress tests, cognitive risk challenges, and governance.
  version: 0.3.0
paths:
...
  /v1/hub/analyze:
    post:
      summary: Analyze Decision Hub
      operationId: analyzeDecisionHub
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HubAnalysisRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HubReport'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
  /v1/hub/reports/{hub_run_id}:
    get:
      summary: Get Decision Hub Report
      operationId: getDecisionHubReport
      parameters:
      - name: hub_run_id
        in: path
        required: true
        schema:
          type: string
          title: Hub Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HubReport'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
  /v1/decisions/evaluate:
    post:
      summary: Evaluate Decision
      operationId: evaluateDecision
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionRequest-Input'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionRecord'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
  /v1/decisions/{decision_id}:
    get:
      summary: Get Decision
      operationId: getDecision
      parameters:
      - name: decision_id
        in: path
        required: true
        schema:
          type: string
          title: Decision Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionRecord'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
  /v1/decisions/{decision_id}/history:
    get:
      summary: Get Decision History
      operationId: getDecisionHistory
      parameters:
      - name: decision_id
        in: path
        required: true
        schema:
          type: string
          title: Decision Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DecisionRecord'
                title: Response Getdecisionhistory
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
  /v1/decisions/{decision_id}/approval:
    post:
      summary: Record Decision Approval
      operationId: recordDecisionApproval
      parameters:
      - name: decision_id
        in: path
        required: true
        schema:
          type: string
          title: Decision Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionRecord'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - bearerAuth: []
components:
  schemas:
    AlgorithmAuditEvent:
      properties:
        sequence:
          type: integer
          minimum: 1.0
          title: Sequence
        stage:
          type: string
          minLength: 1
          title: Stage
        rule_id:
          type: string
          minLength: 1
          title: Rule Id
        operation:
          type: string
          minLength: 1
          title: Operation
        inputs:
          additionalProperties: true
          type: object
          title: Inputs
        output:
          title: Output
        references:
          items:
            type: string
          type: array
          title: References
        previous_event_hash:
          anyOf:
          - type: string
            pattern: ^[0-9a-f]{64}$
          - type: 'null'
          title: Previous Event Hash
        event_hash:
          type: string
          pattern: ^[0-9a-f]{64}$
          title: Event Hash
      additionalProperties: false
      type: object
      required:
      - sequence
      - stage
      - rule_id
      - operation
      - event_hash
      title: AlgorithmAuditEvent
    Alternative:
      properties:
        id:
          type: string
          pattern: ^S[0-9A-Za-z_-]+$
          title: Id
        name:
          type: string
          minLength: 1
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        metrics:
          additionalProperties: true
          type: object
          title: Metrics
        required_assumptions:
          items:
            type: string
          type: array
          title: Required Assumptions
        evidence_ids:
          items:
            type: string
          type: array
          title: Evidence Ids
      additionalProperties: false
      type: object
      required:
      - id
      - name
      title: Alternative
...
    ApprovalRecord:
      properties:
        approved:
          type: boolean
          title: Approved
        approver:
          type: string
          title: Approver
        authorization_ref:
          type: string
          title: Authorization Ref
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        approved_at:
          type: string
          format: date-time
          title: Approved At
      additionalProperties: false
      type: object
      required:
      - approved
      - approver
      - authorization_ref
      title: ApprovalRecord
    ApprovalRequest:
      properties:
        approved:
          type: boolean
          title: Approved
        approver:
          type: string
          minLength: 1
          title: Approver
        authorization_ref:
          type: string
          minLength: 1
          title: Authorization Ref
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      additionalProperties: false
      type: object
      required:
      - approved
      - approver
      - authorization_ref
      title: ApprovalRequest
    Assumption:
      properties:
        id:
          type: string
          pattern: ^A[0-9A-Za-z_-]+$
          title: Id
        statement:
          type: string
          minLength: 1
          title: Statement
        source:
          $ref: '#/components/schemas/AssumptionSource'
        falsification_condition:
          type: string
          minLength: 1
          title: Falsification Condition
        critical:
          type: boolean
          title: Critical
          default: true
        dependencies:
          items:
            type: string
          type: array
          title: Dependencies
        evidence_ids:
          items:
            type: string
          type: array
          title: Evidence Ids
        responsibility:
          anyOf:
          - $ref: '#/components/schemas/ResponsibilityRef'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - id
      - statement
      - source
      - falsification_condition
      title: Assumption
...
    AuditIssue:
      properties:
        code:
          type: string
          title: Code
        severity:
          $ref: '#/components/schemas/IssueSeverity'
        path:
          type: string
          title: Path
        message:
          type: string
          title: Message
        blocking:
          type: boolean
          title: Blocking
          default: false
      additionalProperties: false
      type: object
      required:
      - code
      - severity
      - path
      - message
      title: AuditIssue
    CognitiveAuditReport:
      properties:
        scanner_id:
          type: string
          title: Scanner Id
        version:
          type: string
          title: Version
        findings:
          items:
            $ref: '#/components/schemas/CognitiveRiskFinding'
          type: array
          title: Findings
        counts:
          additionalProperties:
            type: integer
          propertyNames:
            $ref: '#/components/schemas/IssueSeverity'
          type: object
          title: Counts
      additionalProperties: false
      type: object
      required:
      - scanner_id
      - version
      title: CognitiveAuditReport
    CognitiveRiskFinding:
      properties:
        code:
          type: string
          title: Code
        severity:
          $ref: '#/components/schemas/IssueSeverity'
        title:
          type: string
          title: Title
        explanation:
          type: string
          title: Explanation
        evidence_refs:
          items:
            type: string
          type: array
          title: Evidence Refs
        challenge_question:
          type: string
          title: Challenge Question
      additionalProperties: false
      type: object
      required:
      - code
      - severity
      - title
      - explanation
      - challenge_question
      title: CognitiveRiskFinding
    Constraint-Input:
      properties:
        id:
          type: string
          pattern: ^C[0-9A-Za-z_-]+$
          title: Id
        name:
          type: string
          minLength: 1
          title: Name
        kind:
          $ref: '#/components/schemas/ConstraintKind'
        metric:
          type: string
          minLength: 1
          title: Metric
        operator:
          $ref: '#/components/schemas/ConstraintOperator'
        expected:
          title: Expected
        penalty:
          anyOf:
          - type: number
            maximum: 1.0
            exclusiveMinimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Penalty
          description: Explicit score deduction for a failed soft constraint.
        responsibility:
          $ref: '#/components/schemas/ResponsibilityRef'
      additionalProperties: false
      type: object
      required:
      - id
      - name
      - kind
      - metric
      - operator
      - expected
      - responsibility
      title: Constraint
...
    ConstraintKind:
      type: string
      enum:
      - hard
      - soft
      title: ConstraintKind
...
    Criterion-Input:
      properties:
        id:
          type: string
          pattern: ^K[0-9A-Za-z_-]+$
          title: Id
        name:
          type: string
          minLength: 1
          title: Name
        metric:
          type: string
          minLength: 1
          title: Metric
        weight:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Weight
        scoring_rule:
          $ref: '#/components/schemas/ScoringRule'
        min_value:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Min Value
        max_value:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Max Value
        target_value:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Target Value
        responsibility:
          $ref: '#/components/schemas/ResponsibilityRef'
      additionalProperties: false
      type: object
      required:
      - id
      - name
      - metric
      - weight
      - scoring_rule
      - min_value
      - max_value
      - responsibility
      title: Criterion
...
    DecisionRecord:
      properties:
        request:
          $ref: '#/components/schemas/DecisionRequest-Output'
        result:
          $ref: '#/components/schemas/DecisionResult'
        approval:
          anyOf:
          - $ref: '#/components/schemas/ApprovalRecord'
          - type: 'null'
        revision:
          type: integer
          minimum: 1.0
          title: Revision
          default: 1
        parent_record_hash:
          anyOf:
          - type: string
            pattern: ^[0-9a-f]{64}$
          - type: 'null'
          title: Parent Record Hash
        record_hash:
          type: string
          pattern: ^$|^[0-9a-f]{64}$
          title: Record Hash
          default: ''
      additionalProperties: false
      type: object
      required:
      - request
      - result
      title: DecisionRecord
    DecisionRequest-Input:
      properties:
        decision_id:
          anyOf:
          - type: string
            pattern: ^DEC-[0-9A-Za-z_-]+$
          - type: 'null'
          title: Decision Id
        objective:
          type: string
          minLength: 1
          title: Objective
        decision_owner:
          $ref: '#/components/schemas/ResponsibilityRef'
        time_horizon:
          anyOf:
          - type: string
          - type: 'null'
          title: Time Horizon
        evaluation_as_of:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Evaluation As Of
          description: Explicit replay time used for temporal evidence checks.
        evaluation_mode:
          $ref: '#/components/schemas/EvaluationMode'
          default: constraint_only
        criteria:
          items:
            $ref: '#/components/schemas/Criterion-Input'
          type: array
          title: Criteria
        constraints:
          items:
            $ref: '#/components/schemas/Constraint-Input'
          type: array
          title: Constraints
        assumptions:
          items:
            $ref: '#/components/schemas/Assumption'
          type: array
          title: Assumptions
        alternatives:
          items:
            $ref: '#/components/schemas/Alternative'
          type: array
          minItems: 1
          title: Alternatives
        evidence:
          items:
            $ref: '#/components/schemas/Evidence-Input'
          type: array
          title: Evidence
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
      additionalProperties: false
      type: object
      required:
      - objective
      - decision_owner
      - alternatives
      title: DecisionRequest
    DecisionRequest-Output:
      properties:
        decision_id:
          anyOf:
          - type: string
            pattern: ^DEC-[0-9A-Za-z_-]+$
          - type: 'null'
          title: Decision Id
        objective:
          type: string
          minLength: 1
          title: Objective
        decision_owner:
          $ref: '#/components/schemas/ResponsibilityRef'
        time_horizon:
          anyOf:
          - type: string
          - type: 'null'
          title: Time Horizon
        evaluation_as_of:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Evaluation As Of
          description: Explicit replay time used for temporal evidence checks.
        evaluation_mode:
          $ref: '#/components/schemas/EvaluationMode'
          default: constraint_only
        criteria:
          items:
            $ref: '#/components/schemas/Criterion-Output'
          type: array
          title: Criteria
        constraints:
          items:
            $ref: '#/components/schemas/Constraint-Output'
          type: array
          title: Constraints
        assumptions:
          items:
            $ref: '#/components/schemas/Assumption'
          type: array
          title: Assumptions
        alternatives:
          items:
            $ref: '#/components/schemas/Alternative'
          type: array
          minItems: 1
          title: Alternatives
        evidence:
          items:
            $ref: '#/components/schemas/Evidence-Output'
          type: array
          title: Evidence
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
      additionalProperties: false
      type: object
      required:
      - objective
      - decision_owner
      - alternatives
      title: DecisionRequest
    DecisionResult:
      properties:
        decision_id:
          type: string
          title: Decision Id
        status:
          $ref: '#/components/schemas/DecisionStatus'
        objective:
          type: string
          title: Objective
        evaluation_mode:
          $ref: '#/components/schemas/EvaluationMode'
        audit_passed:
          type: boolean
          title: Audit Passed
        issues:
          items:
            $ref: '#/components/schemas/AuditIssue'
          type: array
          title: Issues
        alternatives:
          items:
            $ref: '#/components/schemas/AlternativeEvaluation'
          type: array
          title: Alternatives
        eligible_alternative_ids:
          items:
            type: string
          type: array
          title: Eligible Alternative Ids
        leading_candidate_ids:
          items:
            type: string
          type: array
          title: Leading Candidate Ids
        failure_branches:
          items:
            $ref: '#/components/schemas/FailureBranch'
          type: array
          title: Failure Branches
        responsibility_map:
          items:
            $ref: '#/components/schemas/ResponsibilityEntry'
          type: array
          title: Responsibility Map
        unresolved_variables:
          items:
            type: string
          type: array
          title: Unresolved Variables
        robustness:
          $ref: '#/components/schemas/RobustnessReport'
        counterfactuals:
          items:
            $ref: '#/components/schemas/CounterfactualResult'
          type: array
          title: Counterfactuals
        trace:
          items:
            $ref: '#/components/schemas/TraceEntry'
          type: array
          title: Trace
        algorithm_audit:
          items:
            $ref: '#/components/schemas/AlgorithmAuditEvent'
          type: array
          title: Algorithm Audit
        algorithm_audit_root_hash:
          anyOf:
          - type: string
            pattern: ^[0-9a-f]{64}$
          - type: 'null'
          title: Algorithm Audit Root Hash
        policy:
          $ref: '#/components/schemas/PolicySnapshot'
        engine_version:
          type: string
          title: Engine Version
        input_fingerprint:
          type: string
          pattern: ^[0-9a-f]{64}$
          title: Input Fingerprint
        evaluation_as_of:
          type: string
          format: date-time
          title: Evaluation As Of
        human_approval_required:
          type: boolean
          title: Human Approval Required
          default: true
        generated_at:
          type: string
          format: date-time
          title: Generated At
      additionalProperties: false
      type: object
      required:
      - decision_id
      - status
      - objective
      - evaluation_mode
      - audit_passed
      - issues
      - alternatives
      - eligible_alternative_ids
      - leading_candidate_ids
      - failure_branches
      - responsibility_map
      - unresolved_variables
      - policy
      - engine_version
      - input_fingerprint
      - evaluation_as_of
      title: DecisionResult
    DecisionStatus:
      type: string
      enum:
      - DRAFT
      - STRUCTURED
      - EVIDENCE_PENDING
      - EVALUATED
      - AUDIT_FAILED
      - AUDIT_PASSED
      - HUMAN_APPROVAL_REQUIRED
      - APPROVED
      - REJECTED
      title: DecisionStatus
    EvaluationMode:
      type: string
      enum:
      - constraint_only
      - weighted
      title: EvaluationMode
    Evidence-Input:
      properties:
        id:
          type: string
          pattern: ^E[0-9A-Za-z_-]+$
          title: Id
        statement:
          type: string
          minLength: 1
          title: Statement
        source:
          type: string
          minLength: 1
          title: Source
        status:
          $ref: '#/components/schemas/EvidenceStatus'
          default: supplied
        observed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Observed At
        valid_until:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Valid Until
        quality:
          anyOf:
          - $ref: '#/components/schemas/EvidenceQuality-Input'
          - type: 'null'
        content_hash:
          anyOf:
          - type: string
            pattern: ^[0-9a-fA-F]{64}$
          - type: 'null'
          title: Content Hash
        responsibility:
          $ref: '#/components/schemas/ResponsibilityRef'
      additionalProperties: false
      type: object
      required:
      - id
      - statement
      - source
      - responsibility
      title: Evidence
...
    EvidenceStatus:
      type: string
      enum:
      - supplied
      - missing
      - disputed
      title: EvidenceStatus
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HubAnalysisRequest:
      properties:
        decision:
          $ref: '#/components/schemas/DecisionRequest-Input'
        scenarios:
          items:
            $ref: '#/components/schemas/ScenarioDefinition'
          type: array
          title: Scenarios
        run_cognitive_audit:
          type: boolean
          title: Run Cognitive Audit
          default: true
      additionalProperties: false
      type: object
      required:
      - decision
      title: HubAnalysisRequest
    HubPolicySnapshot:
      properties:
        policy_id:
          type: string
          title: Policy Id
        version:
          type: string
          title: Version
        weight_concentration_threshold:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Weight Concentration Threshold
        authority_concentration_threshold:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Authority Concentration Threshold
        minimum_eligible_alternatives:
          type: integer
          title: Minimum Eligible Alternatives
      additionalProperties: false
      type: object
      required:
      - policy_id
      - version
      - weight_concentration_threshold
      - authority_concentration_threshold
      - minimum_eligible_alternatives
      title: HubPolicySnapshot
    HubReport:
      properties:
        hub_run_id:
          type: string
          pattern: ^HUB-[0-9A-Za-z_-]+$
          title: Hub Run Id
        hub_version:
          type: string
          title: Hub Version
        decision_record:
          $ref: '#/components/schemas/DecisionRecord'
        scenarios:
          items:
            $ref: '#/components/schemas/ScenarioResult'
          type: array
          title: Scenarios
        cognitive_audit:
          anyOf:
          - $ref: '#/components/schemas/CognitiveAuditReport'
          - type: 'null'
        information_priorities:
          items:
            $ref: '#/components/schemas/InformationPriority'
          type: array
          title: Information Priorities
        hub_policy:
          $ref: '#/components/schemas/HubPolicySnapshot'
        algorithm_audit_verified:
          type: boolean
          title: Algorithm Audit Verified
        report_hash:
          type: string
          pattern: ^$|^[0-9a-f]{64}$
          title: Report Hash
          default: ''
        generated_at:
          type: string
          format: date-time
          title: Generated At
      additionalProperties: false
      type: object
      required:
      - hub_run_id
      - hub_version
      - decision_record
      - hub_policy
      - algorithm_audit_verified
      title: HubReport
    InformationPriority:
      properties:
        rank:
          type: integer
          minimum: 1.0
          title: Rank
        variable_ref:
          type: string
          title: Variable Ref
        tier:
          $ref: '#/components/schemas/InformationPriorityTier'
        affected_alternative_ids:
          items:
            type: string
          type: array
          title: Affected Alternative Ids
        affected_leading_candidate_ids:
          items:
            type: string
          type: array
          title: Affected Leading Candidate Ids
        reason:
          type: string
          title: Reason
        recommended_action:
          type: string
          title: Recommended Action
      additionalProperties: false
      type: object
      required:
      - rank
      - variable_ref
      - tier
      - reason
      - recommended_action
      title: InformationPriority
    InformationPriorityTier:
      type: string
      enum:
      - blocking
      - leader_exposed
      - structural
      - review
      title: InformationPriorityTier
    IssueSeverity:
      type: string
      enum:
      - info
      - warning
      - error
      title: IssueSeverity
...
    ResponsibilityRef:
      properties:
        owner:
          type: string
          minLength: 1
          title: Owner
          description: Smallest accountable person or organizational role.
        role:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
        source:
          type: string
          minLength: 1
          title: Source
          description: Document, policy, instruction, or record assigning responsibility.
        authorization_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization Ref
      additionalProperties: false
      type: object
      required:
      - owner
      - source
      title: ResponsibilityRef
...
    ScenarioDefinition:
      properties:
        id:
          type: string
          pattern: ^SC[0-9A-Za-z_-]+$
          title: Id
        name:
          type: string
          minLength: 1
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        metric_overrides:
          additionalProperties:
            additionalProperties: true
            type: object
          type: object
          title: Metric Overrides
        failed_assumption_ids:
          items:
            type: string
          type: array
          title: Failed Assumption Ids
        evidence_status_overrides:
          additionalProperties:
            $ref: '#/components/schemas/EvidenceStatus'
          type: object
          title: Evidence Status Overrides
      additionalProperties: false
      type: object
      required:
      - id
      - name
      title: ScenarioDefinition
    ScenarioOutcomeStatus:
      type: string
      enum:
      - evaluated
      - blocked
      - no_viable_alternative
      title: ScenarioOutcomeStatus
    ScenarioResult:
      properties:
        scenario_id:
          type: string
          title: Scenario Id
        name:
          type: string
          title: Name
        outcome_status:
          $ref: '#/components/schemas/ScenarioOutcomeStatus'
        engine_status:
          $ref: '#/components/schemas/DecisionStatus'
        failed_assumption_ids:
          items:
            type: string
          type: array
          title: Failed Assumption Ids
        invalidated_assumption_ids:
          items:
            type: string
          type: array
          title: Invalidated Assumption Ids
        removed_alternative_ids:
          items:
            type: string
          type: array
          title: Removed Alternative Ids
        eligible_alternative_ids:
          items:
            type: string
          type: array
          title: Eligible Alternative Ids
        leading_candidate_ids:
          items:
            type: string
          type: array
          title: Leading Candidate Ids
        baseline_leading_candidate_ids:
          items:
            type: string
          type: array
          title: Baseline Leading Candidate Ids
        decision_changed:
          type: boolean
          title: Decision Changed
        alternative_scores:
          additionalProperties:
            anyOf:
            - type: string
              pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
            - type: 'null'
          type: object
          title: Alternative Scores
        issues:
          items:
            $ref: '#/components/schemas/AuditIssue'
          type: array
          title: Issues
        scenario_fingerprint:
          type: string
          pattern: ^[0-9a-f]{64}$
          title: Scenario Fingerprint
        algorithm_audit:
          items:
            $ref: '#/components/schemas/AlgorithmAuditEvent'
          type: array
          title: Algorithm Audit
        algorithm_audit_root_hash:
          anyOf:
          - type: string
            pattern: ^[0-9a-f]{64}$
          - type: 'null'
          title: Algorithm Audit Root Hash
        algorithm_audit_verified:
          type: boolean
          title: Algorithm Audit Verified
      additionalProperties: false
      type: object
      required:
      - scenario_id
      - name
      - outcome_status
      - engine_status
      - failed_assumption_ids
      - invalidated_assumption_ids
      - removed_alternative_ids
      - eligible_alternative_ids
      - leading_candidate_ids
      - baseline_leading_candidate_ids
      - decision_changed
      - alternative_scores
      - issues
      - scenario_fingerprint
      - algorithm_audit_verified
      title: ScenarioResult
    ScoringRule:
      type: string
      enum:
      - higher_is_better
      - lower_is_better
      - target_is_better
      title: ScoringRule
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
servers:
- url: https://YOUR-DOMAIN.example.com